Add Trid function Add Zakhary function Add classes for caching double values calculated from a single long value which are expensive to calculate Fix error in fitness search that truncated the search for large fitness factors

+++ src/main/java/uk/edu/populationfitness/models/genes/fitness/Search.java
@@ -19,7 +19,7 @@ public class Search extends FitnessRange {
     }
 
     private double centre(){
-        return ((int)((min() + max()) / increment) / 2) * increment;
+        return ((long)((min() + max()) / increment) / 2) * increment;
     }
 
     public Search current(double current){
